enhanc: mais um avança na navegaçao
