+++ geochat/lib/_shared/route/sailor_navigator.dart
@@ -1,9 +1,7 @@
 import 'package:geochat/_shared/route/navigator.dart';
-import 'package:geochat/home/home_bloc.dart';
-import 'package:geochat/home/home_widget.dart';
 import 'package:sailor/sailor.dart';
 
-final sailor = Sailor();
+final Sailor sailor = Sailor();
 
 class SailorNavigator extends Navigator {
   @override
@@ -13,17 +11,6 @@ class SailorNavigator extends Navigator {
 
   @override
   void pop() {
-    // TODO: implement pop
-  }
-}
-
-class Routes {
-  static void createRoutes() {
-    sailor.addRoute(SailorRoute(
-      name: HomeBloc.route,
-      builder: (context, args, params) {
-        return HomeWidget();
-      },
-    ));
+    sailor.navigatorKey.currentState.pop();
   }
 }
